backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
stripe==7.7.0
razorpay==1.4.1
firebase-admin==6.2.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
python-magic==0.4.27
django-storages==1.14.2
boto3==1.33.6 